Task 0 & 3

+++ 0x0E-structures_typedef/dog.h
@@ -13,7 +13,7 @@ typedef struct dog dog_t;
 struct dog
 {
 	char *name;
-	float age;
+	int age;
 	char *owner;
 };
